+++ package.json
@@ -9,8 +9,10 @@
     "connect-mongo": "^2.0.2",
     "express": "^4.16.4",
     "express-session": "^1.15.6",
+    "mailgen": "^2.0.6",
     "mongodb": "^3.1.10",
     "mongoose": "^5.3.13",
+    "nodemailer": "^4.7.0",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "react": "^16.6.3",
