+++ src/js/server/utils/passportUtil.js
@@ -1,33 +1,17 @@
 // IMPORTANT: 
 // This module assumes it is called in a DB session.
-const passport = require('passport');
-const LocalStrategy = require('passport-local');
+const passport = require('./strategies.js');
 const session = require('express-session');
-
-// Mongoose
+const MongoStore = require('connect-mongo')(session);
 const mongoUtil = require('./mongoUtil.js');
 const schema = require('./schema.js');
 const ObjectID = require('mongodb').ObjectID;
-const MongoStore = require('connect-mongo')(session);
 const User = schema.User;
 
 var setup = false;
 
 let setupPassport = function (app) {
   if (setup) return;
-  
-  passport.use(new LocalStrategy(
-    function (usernameOrEmail, password, done) {
-        User.findOne({$or:[{email: usernameOrEmail},{username: usernameOrEmail}]}, (err, user) => {
-        console.log("User " + usernameOrEmail + " attempted to log in.");
-        if (err) { return done(err); }
-        if (!user) { return done(null, false); }
-        if (!user.validPassword(password)) { return done(null, false); }
-        console.log("User " + usernameOrEmail + " logged in.");
-        return done(null, user);
-      });
-    }
-  ));
 
   app.use(session({
     secret: process.env.SESSION_SECRET,
